backend/requirements.txt
# Core dependencies
python-dotenv==1.0.1
typer==0.12.3
pyyaml==6.0.2

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# HTTP and API
requests==2.32.3

# Data processing
pandas==2.2.3

# Date/time utilities
python-dateutil==2.9.0.post0

# Visualization (for Phase 2)
matplotlib==3.9.0
plotly==5.24.0

# AI (for Phase 3)
anthropic==0.39.0
